+++ docs/images/plot_multiple_runs/check_data.py
@@ -2,10 +2,10 @@ from zacrostools.kmc_output import KMCOutput
 
 conditions = 'CH4_3.728e+00#O2_2.683e-07'
 kmc_output = KMCOutput(path=f'./scan_results_POM_1000K_PtHfC/{conditions}',
-                       window_percent=[50, 100], window_type='time')
+                       analysis_range=[50, 100], range_type='time')
 
 kmc_output_ref = KMCOutput(path=f'./scan_results_POM_1000K_HfC/{conditions}',
-                           window_percent=[50, 100], window_type='time')
+                           analysis_range=[50, 100], range_type='time')
 
 print(f"TOF = {kmc_output.tof['H2']}")
 print(f"TOF (ref) = {kmc_output_ref.tof['H2']}")