+++ docs/images/plot_multiple_runs/customise_colorbar.py
@@ -4,7 +4,7 @@ from zacrostools.plot_functions import plot_heatmap
 fig, axs = plt.subplots(1, figsize=(4.3, 3.5))
 
 cp = plot_heatmap(ax=axs, scan_path="./scan_results_POM_1000K_PtHfC", x="pressure_CH4", y="pressure_O2", z="tof",
-                  gas_spec="H2", window_percent=[50, 100], window_type="time", auto_title=True, show_colorbar=False)
+                  gas_spec="H2", analysis_range=[50, 100], range_type="time", auto_title=True, show_colorbar=False)
 
 # Create colorbar and adjust the size of its tick labels
 cbar = plt.colorbar(cp, ax=axs)