+++ docs/images/plot_multiple_runs/customise_singlesubplot.py
@@ -4,7 +4,7 @@ from zacrostools.plot_functions import plot_heatmap
 fig, axs = plt.subplots(1, figsize=(4.3, 3.5))
 
 plot_heatmap(ax=axs, scan_path="./scan_results_POM_1000K_PtHfC", x="pressure_CH4", y="pressure_O2", z="tof",
-                  gas_spec="H2", window_percent=[50, 100], window_type="time", auto_title=True)
+                  gas_spec="H2", analysis_range=[50, 100], range_type="time", auto_title=True)
 
 # Adjust size of axis ticks
 axs.tick_params(axis='both', which='major', labelsize=14)