+++ docs/images/plot_multiple_runs/scan_tof_dif.py
@@ -6,7 +6,7 @@ fig, axs = plt.subplots(1, figsize=(5, 3.5))
 
 plot_heatmap(ax=axs, scan_path="./scan_results_POM_1000K_PtHfC", scan_path_ref="./scan_results_POM_1000K_HfC",
              x="pressure_CH4", y="pressure_O2", z="tof_dif",
-             gas_spec="H2", window_percent=[50, 100], window_type="time", auto_title=True,
+             gas_spec="H2", analysis_range=[50, 100], range_type="time", auto_title=True,
              levels=np.logspace(-3, 4, num=15))
 
 plt.tight_layout()